% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucd_raw_norm.R
\name{ucd_raw_norm}
\alias{ucd_raw_norm}
\title{Random sampling of age-at-death estimations of archaeological skeletal remains with assumed normal probability distribution}
\usage{
ucd_raw_norm(est_min, est_max, n_it = 2000)
}
\arguments{
\item{est_min}{a vector specifying the lower margin of the age-at-death estimate for every individual.}

\item{est_max}{a vector specifying the upper margin of the age-at-death estimate for every individual.}

\item{n_it}{number of iterations to be rendered. Default = 2000.}
}
\description{
This function generates randomly sampled iterations of exact age-at-death estimations of archaeological remains.
}
\examples{
# Creating input dataset containing the estimated minimum and maximum age of 5 individuals,
# 5, 6-10, 9-13, 3-6 and 4-8 years old.

df <- as.data.frame(cbind(c(5,6,9,3,4),c(5,10,13,6,8)))

colnames(df) <- c("min","max")

ucd_raw_norm(df$min, df$max, n_it=10)

       V1        V2        V3        V4        V5        V6        V7        V8        V9       V10
1  5.000000  5.000000  5.000000  5.000000  5.000000  5.000000  5.000000  5.000000  5.000000  5.000000
2  8.815134  8.182409  8.764919  9.154063  7.518555  9.437651  9.500474  7.627906  8.202944  7.001863
3 10.483655 10.492352 10.808653 11.857053 11.927722 10.784344 10.713252 11.295322 11.077121 11.096039
4  4.006476  3.826322  4.192252  4.180640  4.208959  4.949479  4.373795  3.352460  4.267415  4.832211
5  5.263606  5.500124  6.303949  5.185179  6.199918  6.168494  5.828448  5.831341  5.521000  7.093522
# rows = individuals; columns = iterations
}
\keyword{archaeology}
\keyword{osteology;}
\keyword{random}
\keyword{sampling;}
