% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucd_sd_norm.R
\name{ucd_sd_norm}
\alias{ucd_sd_norm}
\title{Table of standard deviation of random samplings of age-at-death estimations of archaeological skeletal remains with assumed normal probability distribution}
\usage{
ucd_sd_norm(est_min, est_max, age_breaks, n_it = 2000)
}
\arguments{
\item{est_min}{a vector specifying the lower margin of the age-at-death estimate for every individual.}

\item{est_max}{a vector specifying the upper margin of the age-at-death estimate for every individual.}

\item{age_breaks}{a vector specifing the age groups to \code{table()} by.}

\item{n_it}{number of iterations to be rendered. Default = 2000.}
}
\description{
This function generates a table of standard deviation of counts by age group and iteration of randomly
sampled iterations of exact age-at-death estimations of archaeological remains. For details on the
random sampling process, see.
}
\examples{
# Creating input dataset containing the estimated minimum and maximum age of 5 individuals,
# 5, 6-10, 9-13, 3-6 and 4-8 years old.

df <- as.data.frame(cbind(c(5,6,9,3,4),c(5,10,13,6,8)))

colnames(df) <- c("min","max")

ucd_sd_norm(df$min, df$max, age_breaks=seq(0,14,2))

   (0,2]     (2,4]     (4,6]     (6,8]    (8,10]   (10,12]   (12,14]
0.0000000 0.3653005 0.6252771 0.7086516 0.5664841 0.3444348 0.2393902

}
\keyword{archaeology}
\keyword{osteology;}
\keyword{random}
\keyword{sampling;}
